Update Error Status of Delete Route

+++ backend/controllers/product.controller.js
@@ -70,6 +70,13 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+    
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({
+            success: false,
+            message: "Product not Found"
+        });
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(201).json({
@@ -77,9 +84,9 @@ export const deleteProduct = async (req, res) => {
             message: "Product Deleted"
         });
     } catch (error) {
-        res.status(404).json({
+        res.status(500).json({
             success: false,
-            message: "Product not Found"
+            message: "Server Error"
         });
     }
 }
